# swagger/sbAPi.yaml
openapi: 3.0.0
info:
  title: SafeBites OpenAPI 3.0
  description: API for managing and researching restaurants
  license:
    name: MIT
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/filippo.bassetti/SafeBitesApi/1.0.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:8000/api/v1
  description: Localhost
paths:
...
  /review/{rest_id}:
    get:
      summary: View all reviews for a restaurant
      description: Retrieves a list of all reviews for the specified restaurant.
      parameters:
      - name: rest_id
        in: path
        description: The ID of the reviewed restaurant.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: user_id
        in: query
        description: The ID of the user who left the review.
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: A collection of reviews (filtered by restaurant and user if
            provided).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        "404":
          description: No reviews found for the specified restaurant.
    put:
      summary: Update a review for a restaurant by a user
      description: Updates a review for the specified restaurant made by the user.
      parameters:
      - name: rest_id
        in: path
        description: The ID of the reviewed restaurant.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: user_id
        in: query
        description: The ID of the user who left the review.
        required: true
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
      responses:
        "200":
          description: Review updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        "404":
          description: Review not found.
        "400":
          description: Invalid input.
    delete:
      summary: Delete a review for a restaurant by a user
      description: Deletes the review for the specified restaurant made by the user.
      parameters:
      - name: rest_id
        in: path
        description: The ID of the reviewed restaurant.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: user_id
        in: query
        description: The ID of the user who left the review.
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "204":
          description: Review deleted successfully (No Content).
        "404":
          description: No matching review found for the specified restaurant and user.
  /review:
    post:
      summary: Create a review for a restaurant
      description: Creates a new review for a restaurant by a user.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
      responses:
        "201":
          description: Review created. The Location header contains a link to the
            newly created review.
          headers:
            Location:
              description: Link to the newly created review.
              style: simple
              explode: false
              schema:
                type: string
components:
  schemas:
...
    Review:
      required:
      - rest_id
      - text
      - user_id
      type: object
      properties:
        rest_id:
          type: string
          description: ID of the reviewed restaurant.
        user_id:
          type: string
          description: ID of the user who left the review.
        text:
          type: string
          description: Text of the review.
...